actionsNav is not zettel specific
